Revert "useSetColumnDefaultSortOrder"
This reverts commit 0ead037c7329a05f5cb4e72da20986545acf93ac.
