+++ src/pages/task/testsTable/TestsTableCore.tsx
@@ -33,7 +33,6 @@ import {
 import { ColumnProps } from "antd/es/table";
 import { Table, Skeleton } from "antd";
 import { isNetworkRequestInFlight } from "apollo-client/core/networkStatus";
-import { useSetColumnDefaultSortOrder } from "hooks/useSetColumnDefaultSortOrder";
 
 const arrayFormat = "comma";
 
@@ -49,11 +48,7 @@ export const TestsTableCore: React.FC = () => {
     id,
     ...getQueryVariables(search),
   });
-  useSetColumnDefaultSortOrder<TestResult>(
-    columns,
-    initialQueryVariables.cat,
-    initialQueryVariables.dir
-  );
+
   const { data, fetchMore, networkStatus, error } = useQuery<
     TaskTestsQuery,
     TaskTestsQueryVariables
@@ -62,6 +57,7 @@ export const TestsTableCore: React.FC = () => {
     notifyOnNetworkStatusChange: true,
   });
   useDisableTableSortersIfLoading(networkStatus);
+
   // this fetch is when url params change (sort direction, sort category, status list)
   // and the page num is set to 0
   useEffect(
@@ -108,8 +104,10 @@ export const TestsTableCore: React.FC = () => {
   };
 
   // initial table sort button state to reflect initial URL query params
-  const { pageNum, limitNum } = getQueryVariables(search);
+  const { cat, dir, pageNum, limitNum } = getQueryVariables(search);
 
+  columns.find(({ key }) => key === cat).defaultSortOrder =
+    dir === SortDirection.Asc ? "ascend" : "descend";
   const isLoading = isNetworkRequestInFlight(networkStatus);
   return (
     <>
